requirements.txt
streamlit
openai-whisper
googletrans==4.0.0-rc1
streamlit-audiorecorder
gTTS